opendocument writer: Produce ODTs with 2 images that libreoffice can open
libreoffice was unable to open odt files produced by pandoc if they
contained more than 1 image.  e.g. the markdown:

    ![](img1.svg)
    ![](img1.svg)

Passed through pandoc to create an ODT causes libreoffice (4.1.3.2) to
produce the an error in a dialog box reading:

> Read-Error.
> Error reading file

Libreoffice doesn't seem to like the values of the `draw:name` attributes
to appear more than once on different `draw:frame` XML elements.  I suspect
if the `draw:name` attribute is missing libreoffice internally defaults
to some fixed value causing a conflict if more than one image is present
in the document.

It seems like this technically a libreoffice bug, but it is important for
pandoc to produce ODT files which libreoffice can read given that it and
its cousin openoffice are likely to be the most common programs to be used
to open ODTs so this commit works around this issue in pandoc.

This commit is probably more useful as a guide as to where the code to be
fixed should be fixed rather than something that can be applied directly
to master.  It is intended as a discussion point.  I do not know Haskell
and have not written any before.  Also a test case is missing.

Fixes #1069